+++ src/services/mod.rs
@@ -1,9 +1,8 @@
 pub mod auth;
 pub mod balances;
-pub mod liquidity_adding;
-pub mod liquidity_removing;
 pub mod merge_conditionals_for_underlying;
 pub mod new_mint;
 pub mod redeem_conditionals;
 pub mod swaps;
 pub mod transactions;
+pub mod liquidity;
